+++ packages/index/src/filter/filesize-filter.js
@@ -1,9 +1,22 @@
-
 const log = require('@home-gallery/logger')('index.filter.maxFilesize')
 
 const { parseFilesize } = require('../utils')
 
-const createFilesizeFilter = (maxFilesize) => {
+const createName2Entry = (entries, maxSize) => {
+  return entries
+    .filter(entry => entry.size > maxSize)
+    .reduce((result, entry) => {
+      result[entry.filename] = entry
+      return result
+    }, {})
+}
+
+const isKnown = (name2entry, filename, stat) => {
+  const entry = name2entry[filename]
+  return entry && entry.size == stat.size
+}
+
+const createFilesizeFilter = (entries, maxFilesize, keepKnownFiles = false) => {
   const maxSize = parseFilesize(maxFilesize)
 
   if (maxSize === false) {
@@ -11,9 +24,16 @@ const createFilesizeFilter = (maxFilesize) => {
   }
 
   log.info(`Limit files up to ${maxFilesize} (${maxSize} bytes)`)
-  return (_, stat) => {
-    const { size } = stat
-    return size <= maxSize
+
+  if (keepKnownFiles) {
+    const name2entry = createName2Entry(entries, maxSize)
+    return (filename, stat) => {
+      return stat.size <= maxSize || isKnown(name2entry, filename, stat)
+    }
+  } else {
+    return (_, stat) => {
+      return stat.size <= maxSize
+    }
   }
 }