+++ packages/index/src/filter/index.js
@@ -6,11 +6,11 @@ const { createLimitFilter } = require('./limit-filter')
 
 const asyncFileFilter = promisify(fileFilter)
 
-const createFilter = async (entryCount, options) => {
+const createFilter = async (entries, options) => {
   const filters = []
 
   if (options.maxFilesize) {
-    filters.push(createFilesizeFilter(options.maxFilesize))
+    filters.push(createFilesizeFilter(entries, options.maxFilesize, options.keepKnownFiles))
   }
 
   if (options.exclude && options.exclude.length || options.excludeFromFile) {
@@ -18,7 +18,7 @@ const createFilter = async (entryCount, options) => {
     filters.push(excludeFilter)
   }
 
-  return createLimitFilter(entryCount, options.addLimits, createFilterChain(filters))
+  return createLimitFilter(entries.length, options.addLimits, createFilterChain(filters))
 }
 
 module.exports = {